Conditionals, API to rewrite text
Interface to allow visible text to be rewritten.
Slightly better display of conditionally required values, where the
indicator is only shown if the document requires it. Swaps showing an
indicator when it might be required (but may not be) for not showing an
indicator when it might be. Would require complex client side stuff to
resolve properly, but this way might be better.
Update for newer JRuby.
Update company name.
fossil: 9f5ebb0abf